apollo client config pushed to its folder with all configuration
